==> analysis/02_EDA.Rmd <==
---
title: "Exploratory data analysis - proteomics data"
author: "Emily Johnson"
date: "2022-09-28"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Load libraries

```{r, message=FALSE, warning=FALSE}

library(tidyverse)
library(cowplot) # plot theme
library(mixOmics) # multi-level PCA
library(PCAtools) # pca & correlation plot
library(viridis) # colour schemes
library(ggpubr) # utilities to arrange/annotate ggplot2 plots

```


## Load data

```{r}

# Data
sf_data_norm <- readRDS("./data/sf_data_norm.rds")
plasma_data_norm <- readRDS("./data/plasma_data_norm.rds")


# Pheno
sf_pheno <- readRDS("./data/sf_pheno.rds")
plasma_pheno <- readRDS("./data/plasma_pheno.rds")

```


## PCA

### Plasma

```{r}

# Colour scheme for plasma
redblue<-colorRampPalette(c("#db092c", "#984ea3", "#097ddb"))(50)

row.names(plasma_pheno) <- plasma_pheno$Sample.Number
plasma_pca_res <- pca(na.omit(plasma_data_norm), metadata = plasma_pheno, scale = TRUE, center = TRUE)
plasma_df_out <- as.data.frame(plasma_pca_res$rotated)

```


```{r, message=FALSE, warning=FALSE}

use_cex <- 16/16

plasma_corplot <- eigencorplot(plasma_pca_res,
    metavars = c("Day", "Horse", "Date"),
    col = c( "blue2", "blue1", "black", "red1", "red2"),
    colCorval = "white",
    scale = TRUE,
    main = "",
    plotRsquared = FALSE,
    cexTitleX= use_cex,
    cexTitleY= use_cex,
    cexLabX = use_cex,
    cexLabY = use_cex,
    cexMain = use_cex,
    cexLabColKey = use_cex,
    cexCorval = use_cex) 


```

```{r}

plasma_pca_plot <- ggplot(plasma_df_out, aes(x=PC1, y=PC2, colour = plasma_pheno$Day, shape=factor(plasma_pheno$Horse))) + 
    geom_point(size = 4.5) + 
    xlab(paste0('PC1: ', round(as.numeric(plasma_pca_res$variance[1])), '% expl.var')) + 
    ylab(paste0('PC2: ', round(as.numeric(plasma_pca_res$variance[2])), '% expl.var')) + 
    scale_color_gradientn(colours = redblue) +
    theme_cowplot() + 
    scale_shape_manual(values=c(19,15,17,18)) +
    labs(colour = "Day", shape = "Horse")

```

```{r}

plasma_pca_plot2 <- ggplot(plasma_df_out, aes(x=PC1, y=PC3, colour = plasma_pheno$Day, shape=factor(plasma_pheno$Horse))) + 
    geom_point(size = 4.5) + 
    xlab(paste0('PC1: ', round(as.numeric(plasma_pca_res$variance[1])), '% expl.var')) + 
    ylab(paste0('PC3: ', round(as.numeric(plasma_pca_res$variance[3])), '% expl.var')) + 
    scale_color_gradientn(colours = redblue) +
    theme_cowplot() + 
    scale_shape_manual(values=c(19,15,17,18)) +
    labs(colour = "Day", shape = "Horse")

```

```{r}

plasma_pca_plot3 <- ggplot(plasma_df_out, aes(x=PC4, y=PC5, colour = plasma_pheno$Horse)) + 
    geom_point(size = 4.5) + 
    xlab(paste0('PC4: ', round(as.numeric(plasma_pca_res$variance[5])), '% expl.var')) + 
    ylab(paste0('PC5: ', round(as.numeric(plasma_pca_res$variance[6])), '% expl.var')) + 
    scale_colour_brewer(palette = "Set1") +
    theme_cowplot() + 
    labs(colour = "Horse")

```

```{r}

plasma_variance_plot <- screeplot(plasma_pca_res,
                                  components=1:8,
                                  axisLabSize = 18,
                                  title = "") + theme_cowplot() 

```


```{r, echo=FALSE, fig.height = 8, fig.width = 10, fig.align="center"}

ggarrange(plasma_pca_plot, plasma_pca_plot2, plasma_pca_plot3, plasma_variance_plot,
          nrow = 2, 
          ncol = 2,
          labels = c("A", "B", "C", "D"))

```

```{r, echo=FALSE, fig.height = 4, fig.width = 8, fig.align="center"}

ggarrange(plasma_corplot,
          nrow = 1, 
          ncol = 1,
          labels = c("E"))

```


### Synovial fluid

```{r}

row.names(sf_pheno) <- sf_pheno$Sample.Number
#sf_pheno$Group_day <- as.factor(paste0(sf_pheno$Group, "_", sf_pheno$Day))

sf_pca_res <- pca(na.omit(sf_data_norm), metadata = sf_pheno, scale = TRUE, center = TRUE)
sf_df_out <- as.data.frame(sf_pca_res$rotated)

```


```{r, message=FALSE, warning=FALSE}

use_cex <- 16/16

sf_corplot <- eigencorplot(sf_pca_res,
    metavars = c("Day", "Horse", "Date", "Group"),
    col = c( "blue2", "blue1", "black", "red1", "red2"),
    colCorval = "white",
    scale = TRUE,
    main = "",
    plotRsquared = FALSE,
    cexTitleX= use_cex,
    cexTitleY= use_cex,
    cexLabX = use_cex,
    cexLabY = use_cex,
    cexMain = use_cex,
    cexLabColKey = use_cex,
    cexCorval = use_cex) 


```



```{r}

sf_pca_plot <- ggplot(sf_df_out, aes(x=PC1, y=PC2, colour = sf_pheno$Day, shape=factor(sf_pheno$Horse))) + 
  geom_point(size = 4.5) + 
  xlab(paste0('PC1: ', round(as.numeric(sf_pca_res$variance[1])), '% expl.var')) + 
  ylab(paste0('PC2: ', round(as.numeric(sf_pca_res$variance[2])), '% expl.var')) + 
  scale_color_gradientn(colours = hcl.colors(50, "BluYl")) +
  theme_cowplot() + 
  scale_shape_manual(values=c(19,15,17,18)) +
  labs(colour = "Day", shape = "Horse")

```

```{r}

sf_pca_plot2 <- ggplot(sf_df_out, aes(x=PC1, y=PC3, colour = as.factor(sf_pheno$Group))) + 
  geom_point(size = 4.5) + 
  xlab(paste0('PC1: ', round(as.numeric(sf_pca_res$variance[1])), '% expl.var')) + 
  ylab(paste0('PC3: ', round(as.numeric(sf_pca_res$variance[5])), '% expl.var')) + 
  scale_colour_brewer(palette = "Set1") +
  theme_cowplot() + 
  labs(colour = "Group")

```

```{r}

sf_pca_plot3 <- ggplot(sf_df_out, aes(x=PC1, y=PC4, colour = sf_pheno$Horse)) + 
  geom_point(size = 4.5) + 
  xlab(paste0('PC1: ', round(as.numeric(sf_pca_res$variance[1])), '% expl.var')) + 
  ylab(paste0('PC4: ', round(as.numeric(sf_pca_res$variance[4])), '% expl.var')) + 
  scale_colour_brewer(palette = "Set1") +
  theme_cowplot() + 
  labs(colour = "Horse")

```

```{r}

sf_variance_plot <- screeplot(sf_pca_res,
                                  components=1:8,
                                  axisLabSize = 18,
                                  title = "") + theme_cowplot() 

```


```{r, echo=FALSE, fig.height = 8, fig.width = 10, fig.align="center"}

ggarrange(sf_pca_plot, sf_pca_plot2, sf_pca_plot3, sf_variance_plot,
          nrow = 2, 
          ncol = 2,
          labels = c("A", "B", "C", "D"))

```

```{r, echo=FALSE, fig.height = 4, fig.width = 8, fig.align="center"}

ggarrange(sf_corplot,
          nrow = 1, 
          ncol = 1,
          labels = c("E"))

```

## Multi-level PCA

### Plasma

```{r, fig.show='hide', message=FALSE}

plasma_pca_multi_level <- mixOmics::pca(t(na.omit(plasma_data_norm)), scale = TRUE, center = TRUE, ncomp = 10,
                            multilevel = plasma_pheno$Horse) %>%
                            plotIndiv(., group = plasma_pheno$Day, 
                            ind.names = plasma_pheno$Horse,
                            legend = TRUE, legend.title = 'Day') 

```

```{r}

plasma_ml_pca_plot <- ggplot(plasma_pca_multi_level$df, aes(x=x, y=y, colour = plasma_pheno$Day)) + 
    geom_point(size = 4.5) + 
    xlab(plasma_pca_multi_level[["graph"]][["labels"]]$x) + 
    ylab(plasma_pca_multi_level[["graph"]][["labels"]]$y) + 
    scale_color_gradientn(colours = redblue) +
    theme_cowplot() + 
    labs(colour = "Treatment", shape = "Day") + 
    ggtitle("Plasma multi-level PCA")


```

```{r, echo=FALSE}

plasma_ml_pca_plot

```


### Synovial fluid 

First run the analysis with the control samples still present: 

```{r, fig.show='hide', message=FALSE}

sf_pca_multi_level <- mixOmics::pca(t(na.omit(sf_data_norm)), scale = TRUE, center = TRUE, ncomp = 10,
                            multilevel = sf_pheno$Horse) %>%
                            plotIndiv(., group = sf_pheno$Day, 
                            ind.names = sf_pheno$Horse,
                            legend = TRUE, legend.title = 'Day') 

```

```{r}

sf_ml_pca_plot <- ggplot(sf_pca_multi_level$df, aes(x=x, y=y, colour = sf_pheno$Day)) + 
    geom_point(size = 4.5) + 
    xlab(sf_pca_multi_level[["graph"]][["labels"]]$x) + 
    ylab(sf_pca_multi_level[["graph"]][["labels"]]$y) + 
    scale_color_gradientn(colours = hcl.colors(50, "BluYl")) +
    theme_cowplot() + 
    labs(colour = "Treatment", shape = "Day") + 
    ggtitle("Synovial fluid multi-level PCA")



sf_ml_pca_plot2 <- ggplot(sf_pca_multi_level$df, aes(x=x, y=y, colour = sf_pheno$Group)) + 
    geom_point(size = 4.5) + 
    xlab(sf_pca_multi_level[["graph"]][["labels"]]$x) + 
    ylab(sf_pca_multi_level[["graph"]][["labels"]]$y) + 
    scale_colour_brewer(palette = "Set1") +
    theme_cowplot() + 
    labs(colour = "Day") 


```

```{r, echo=FALSE, fig.height = 4, fig.width = 9, fig.align="center"}

ggarrange(sf_ml_pca_plot, sf_ml_pca_plot2,
          ncol = 2,
          nrow = 1, 
          labels = c("A", "B")) 

```


Now run without the control samples since the DE analysis will be ran vs time 0 instead of against the control:

```{r}

sf_pheno_OA <- sf_pheno[!sf_pheno$Group == "Control",]
sf_data_norm_just_OA <- sf_data_norm[,sf_pheno_OA$Sample.Number]

```


```{r, fig.show='hide', message=FALSE}

sf_pca_multi_level_OA <- mixOmics::pca(t(na.omit(sf_data_norm_just_OA)), scale = TRUE, center = TRUE, ncomp = 10,
                            multilevel = sf_pheno_OA$Horse) %>%
                            plotIndiv(., group = sf_pheno_OA$Day, 
                            ind.names = sf_pheno_OA$Horse,
                            legend = TRUE, legend.title = 'Day',
                            title = '(B) Multilevel PCA of proteomics data') 

```

```{r}

sf_ml_pca_plot_OA <- ggplot(sf_pca_multi_level_OA$df, aes(x=x, y=y, colour = sf_pheno_OA$Day)) + 
    geom_point(size = 4.5) + 
    xlab(sf_pca_multi_level_OA[["graph"]][["labels"]]$x) + 
    ylab(sf_pca_multi_level_OA[["graph"]][["labels"]]$y) + 
    scale_color_gradientn(colours = hcl.colors(50, "BluYl")) +
    theme_cowplot() + 
    labs(colour = "Day") + 
    ggtitle("Synovial fluid multi-level PCA - no control samples")

```

```{r, echo=FALSE}

sf_ml_pca_plot_OA

```
