==> analysis/06_jDR.Rmd <==
---
title: "Multi-omics - joint dimensionality reduction"
author: "Emily Johnson"
date: "2022-09-28"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
bibliography: references.bib
---

# Introduction

This workflow employs joint dimensionality reduction to identify factors (combinations of all the different features) that account for the majority of the variation in the -omics data. These factors can be correlated with the phenotype data for downstream analysis. We can also see how much each -omics contributes to each factor. 

In this analysis the different types of small RNAs were read in individually and treated as different views. However, a more conventional version of the analysis with all the sequencing data collapsed together was also carried out [here](06_jDR_2.html). 


# Load libraries

```{r, message=FALSE, warning=FALSE}

source("./code/utility-functions.R") # load helper functions 
source("./code/ArchRpalettes.R") # additional colour palettes
library(tidyverse)
library(MOFA2) # factor analysis
library(ggpubr) # utilities to arrange/annotate ggplot2 plots
library(cowplot) # plot theme
library(viridis) # colour schemes

```

# Load and process data

```{r}

# Plasma 
plasma_annotation <- readRDS("./data/plasma_proteomics_annotation.rds")
plasma_pheno <- readRDS("./data/plasma_pheno.rds") 
plasma_proteomics_data <- readRDS("./data/plasma_data_norm.rds")
plasma_miRNA_data <- readRDS("./data/plasma_miRNA_data.rds") 
plasma_lncRNA_data <- readRDS("./data/plasma_lncRNA_data.rds")
plasma_snRNA_data <- readRDS("./data/plasma_snRNA_data.rds")

# SF
sf_annotation <- readRDS("./data/sf_proteomics_annotation.rds")
sf_pheno <- readRDS("./data/sf_pheno.rds")
sf_proteomics_data <- readRDS("./data/sf_data_norm.rds")
sf_miRNA_data <- readRDS("./data/sf_miRNA_data.rds")
sf_lncRNA_data <- readRDS("./data/sf_lncRNA_data.rds")
sf_snRNA_data <- readRDS("./data/sf_snRNA_data.rds")


```

Convert 0s to NAs as previous analysis showed they were sampling zeroes rather than biological zeroes:

```{r}

# Plasma 
plasma_miRNA_data[plasma_miRNA_data == 0] <- NA
plasma_lncRNA_data[plasma_lncRNA_data == 0] <- NA
plasma_snRNA_data[plasma_snRNA_data == 0] <- NA

# SF
sf_miRNA_data[sf_miRNA_data == 0] <- NA
sf_lncRNA_data[sf_lncRNA_data == 0] <- NA
sf_snRNA_data[sf_snRNA_data == 0] <- NA


```

# MOFA

In this analysis MOFA2 was used to carry out joint dimensionality reduction. MOFA2 is a widely used package for integrating multi-omics data. It has previously been employed to study arthritis in mice in a similar study (in terms of sample size) [@li2022].

## Plasma

Visualuse the distributions of the data:

```{r, echo=FALSE, message=FALSE, warning=FALSE}

proteomics_hist <- gg_histogram(plasma_proteomics_data, ylab="Frequency", xlab="Log2 Intensity")
miRNA_hist <- gg_histogram(plasma_miRNA_data, ylab="Frequency", xlab="Log2 CPM")
lncRNA_hist <- gg_histogram(plasma_lncRNA_data, ylab="Frequency", xlab="Log2 CPM")
snRNA_hist <- gg_histogram(plasma_snRNA_data, ylab="Frequency", xlab="Log2 CPM")


ggarrange(proteomics_hist, miRNA_hist, lncRNA_hist, snRNA_hist)

```

MOFA2 needs a column in the phenotype data called 'sample'. Rename the 'Sample_ID' column to 'sample'.

```{r}

colnames(plasma_pheno)[5] <- "sample"

```

As an input MOFA2 will take a list containing all the different data. Create this list:

```{r}

plasma_mofa_dat <- list(proteomics = as.matrix(plasma_proteomics_data),
                 miRNA = as.matrix(plasma_miRNA_data),
                 lncRNA = as.matrix(plasma_lncRNA_data),
                 snRNA = as.matrix(plasma_snRNA_data))

```

Create MOFAobject:

```{r}

MOFAobject <- create_mofa(plasma_mofa_dat)
print(MOFAobject)

```

The number of features per view can be visualised with `plot_data_overview`. Missing samples would be represented as gaps in the block colours.

```{r}

plot_data_overview(MOFAobject)

```

Set model parameters. Here three objects are created containing the data options, model options and training options. If the raw counts were used a poisson distribution would be more appropriate.

```{r}

# Scale views
data_opts <- get_default_data_options(MOFAobject)
data_opts$scale_views <- TRUE
head(data_opts)


# Set number of factors
model_opts <- get_default_model_options(MOFAobject)
model_opts$num_factors <- 6
#model_opts$likelihoods[2:4] <- "poisson"
head(model_opts)


# Set training options
train_opts <- get_default_training_options(MOFAobject)
train_opts$convergence_mode <- "slow"
train_opts$seed <- 123
train_opts$maxiter <- 3000
head(train_opts)

```

Prepare MOFA object with modified parameters:

```{r}

MOFAobject <- prepare_mofa(
  object = MOFAobject,
  data_options = data_opts,
  model_options = model_opts,
  training_options = train_opts
)

```

Train model:

```{r, message=FALSE, warning=FALSE}

outfile <- file.path(getwd(),"plasma_model.hdf5")
MOFAobject.trained <- run_mofa(MOFAobject, outfile, use_basilisk=TRUE)

```

Add phenotype data to the model:

```{r}

samples_metadata(MOFAobject.trained) <- plasma_pheno

```

Visualise variance explained:

```{r}

plot_variance_explained(
  MOFAobject.trained, 
  x = "view", 
  y = "factor", 
  plot_total = T
)

```

Majority is from proteomics data - likely due to the higher number of features.

Plot correlation between factors:

```{r}

plot_factor_cor(MOFAobject.trained)

```

Correlation between factors and metadata:

```{r, message=FALSE, warning=FALSE}

correlate_factors_with_covariates(MOFAobject.trained, 
  covariates = c("Day", "Horse"), 
  plot="log_pval", 
  alpha="0.15" # p-value threshold
)


correlate_factors_with_covariates(MOFAobject.trained, 
  covariates = c("Day", "Horse"), 
  plot="r"
)


```

Factor 2 is significantly correlated with "day". Factors 5 and 6 are correlated with horse. Most of the variance in factor 2 is explained by the proteomics data. The variance in factor 1 mostly comes from the miRNA and lncRNA data, however that doesn't appear to correspond to a biological effect. Factor 3 shows slight correlation with "day". As the variance is explained by the proteomics and snRNA data this might also be interesting to explore.

We can inspect the weights of the different features contributing to each factor.

```{r}

proteomics_f2 <- plot_top_weights(MOFAobject.trained, 
                  view = "proteomics", 
                  factor = 2, 
                  nfeatures = 10,
                  scale = F, 
                  abs = T
)

# Annotate so Uniprot accessions are replaced by protein names
# The accessions are stored as factors so need to be converted back into character vectors
proteomics_f2$data$feature_id <- plasma_annotation[as.character(proteomics_f2$data$feature_id),]

```

```{r, echo=FALSE}

proteomics_f2

```

Plot weights for factor 2:

```{r}

plot_weights(MOFAobject.trained,
 view = "proteomics",
 factor = 2,
 nfeatures = 50,     # Top number of features to highlight
 scale = F           # Scale weights from -1 to 1
)


plot_weights(MOFAobject.trained,
 view = "miRNA",
 factor = 2,
 nfeatures = 25,     # Top number of features to highlight
 scale = F           # Scale weights from -1 to 1
)


```

The distribution of weights for the top ranked proteins contributing towards factor 2 are heavily negatively skewed. Only Q28377, F7DZE7, A0A5F5PEG5, and A0A5F5PWX5 are positively associated in the top 50. If the weights for miRNA are plotted it can be seen their contribution to factor 2 is low overall, as the weights range from -0.001 to 0.001.

If we plot factor 1 instead, the weight of the miRNAs is much higher.

```{r}

miRNA_f2 <- plot_top_weights(MOFAobject.trained, 
                  view = "miRNA", 
                  factor = 2, 
                  nfeatures = 10,
                  scale = F, 
                  abs = T
)

miRNA_f1 <- plot_top_weights(MOFAobject.trained, 
                  view = "miRNA", 
                  factor = 1, 
                  nfeatures = 10,
                  scale = F, 
                  abs = T
)

```

```{r, echo=FALSE}

ggarrange(miRNA_f2, miRNA_f1)

```

It is interesting that miR-126-5P and miR-126-3P are among the miRNAs contributing to factor 2. <https://pubmed.ncbi.nlm.nih.gov/33627637/>

Visualise the weights for the proteomics and snRNA data contributing to factor 3:

```{r}

snRNA_f3 <- plot_top_weights(MOFAobject.trained, 
                  view = "snRNA", 
                  factor = 3, 
                  nfeatures = 10,
                  scale = F, 
                  abs = T
)


snRNA_f4 <- plot_top_weights(MOFAobject.trained, 
                  view = "snRNA", 
                  factor = 4, 
                  nfeatures = 10,
                  scale = F, 
                  abs = T
)

```

```{r, echo=FALSE}

ggarrange(snRNA_f3, snRNA_f4)

```

```{r}

proteomics_f3 <- plot_top_weights(MOFAobject.trained, 
                  view = "proteomics", 
                  factor = 3, 
                  nfeatures = 10,
                  scale = F, 
                  abs = T
)

# Annotate so Uniprot accessions are replaced by protein names
# The accessions are stored as factors so need to be converted back into character vectors
proteomics_f3$data$feature_id <- plasma_annotation[as.character(proteomics_f3$data$feature_id),]

```

```{r, echo=FALSE}

proteomics_f3

```

Create heatmaps for proteomics and miRNA data (factor 2):

```{r}

# Add day as a factor to metadata for visualisation purposes
MOFAobject.trained@samples_metadata$Days <- as.factor(plasma_pheno$Day)

# Palette for heatmap
heatmap_cols <- colorRampPalette(ArchRPalettes$coolwarm)(50)

# Plot heatmap
f2_plasma_protein_heatmap <- plot_data_heatmap(MOFAobject.trained, 
  view = "proteomics",
  factor = 2,  
  features = 25,
  cluster_rows = TRUE, cluster_cols = TRUE,
  show_rownames = TRUE, show_colnames = FALSE,
  scale = "row",
  denoise = TRUE,
  color = heatmap_cols,
  border_color = NA,
  annotation_samples = c("Day", "Horse"),
  fontsize = 12
)


f2_plasma_protein_heatmap 

```

```{r}

# color = met.brewer(name="Hiroshige", n=123),
# color = hcl.colors(50, "RdYlBu"),

f2_plasma_miRNA_heatmap <- plot_data_heatmap(MOFAobject.trained, 
  view = "miRNA",
  factor = 2,  
  features = 25,
  cluster_rows = TRUE, cluster_cols = TRUE,
  show_rownames = TRUE, show_colnames = FALSE,
  scale = "row",
  denoise = TRUE,
  color = heatmap_cols,
  border_color = NA,
  annotation_samples = c("Day", "Horse"),
  fontsize = 12
)

f2_plasma_miRNA_heatmap

```

Project the samples onto factor 2 and 3 for each time point.

```{r}

plot_factor(MOFAobject.trained, 
  factors = 2, 
  color_by = "Days",
  add_violin = TRUE,
  dodge = TRUE
)


plot_factor(MOFAobject.trained, 
  factors = 3, 
  color_by = "Days",
  add_violin = TRUE,
  dodge = TRUE
)

```

```{r}

# To extract per sample factor weights manually:
plasma_factors <- get_factors(
    MOFAobject.trained,
    groups = "group1",
    factors = "all",
    scale = FALSE,
    as.data.frame = FALSE
)

```

Save results:

```{r}

plasma_res <- get_weights(
    MOFAobject.trained,
    views = "all",
    factors = "all",
    abs = FALSE,
    scale = FALSE,
    as.data.frame = FALSE
)



plasma_res_df <- get_weights(
    MOFAobject.trained,
    views = "all",
    factors = "all",
    abs = FALSE,
    scale = FALSE,
    as.data.frame = TRUE
)




```

```{r}

saveRDS(plasma_res, "./data/mofa2_plasma.rds")
saveRDS(plasma_res_df, "./data/mofa2_plasma_df.rds")

```

## Synovial fluid - version 1 (MOFA2)

```{r, echo=FALSE, message=FALSE, warning=FALSE}

proteomics_hist <- gg_histogram(sf_proteomics_data, ylab="Frequency", xlab="Log2 Intensity")
miRNA_hist <- gg_histogram(sf_miRNA_data, ylab="Frequency", xlab="Log2 CPM")
lncRNA_hist <- gg_histogram(sf_lncRNA_data, ylab="Frequency", xlab="Log2 CPM")
snRNA_hist <- gg_histogram(sf_snRNA_data, ylab="Frequency", xlab="Log2 CPM")


ggarrange(proteomics_hist, miRNA_hist, lncRNA_hist, snRNA_hist)

```

```{r}

colnames(sf_pheno)[6] <- "sample"

```

```{r}

sf_mofa_dat <- list(proteomics = as.matrix(sf_proteomics_data),
                        miRNA = as.matrix(sf_miRNA_data),
                        lncRNA = as.matrix(sf_lncRNA_data),
                        snRNA = as.matrix(sf_snRNA_data))

```

Create MOFAojbect:

```{r}

MOFAobject <- create_mofa(sf_mofa_dat)
print(MOFAobject)

```

```{r}

plot_data_overview(MOFAobject)

```

Modify model parameters:

```{r}

# Scale views
data_opts <- get_default_data_options(MOFAobject)
data_opts$scale_views <- TRUE
head(data_opts)


# Set number of factors
model_opts <- get_default_model_options(MOFAobject)
model_opts$num_factors <- 6
head(model_opts)


# Set training options
train_opts <- get_default_training_options(MOFAobject)
train_opts$convergence_mode <- "slow"
train_opts$seed <- 123
train_opts$maxiter <- 3000
head(train_opts)

```

Prepare MOFA object with modified parameters:

```{r}

MOFAobject <- prepare_mofa(
  object = MOFAobject,
  data_options = data_opts,
  model_options = model_opts,
  training_options = train_opts
)

```

Train model:

```{r, message=FALSE, warning=FALSE}

outfile <- file.path(getwd(),"sf_model.hdf5")
MOFAobject.trained <- run_mofa(MOFAobject, outfile, use_basilisk=TRUE)

```

Add phenotype data to the model:

```{r}

samples_metadata(MOFAobject.trained) <- sf_pheno

# Add a group-by-day factor 
MOFAobject.trained@samples_metadata$Group_day <- as.factor(paste0(sf_pheno$Group, "_", sf_pheno$Day))

```

Visualise variance explained:

```{r}

plot_variance_explained(
  MOFAobject.trained, 
  x = "view", 
  y = "factor", 
  plot_total = T
)


```

Majority is from proteomics data - likely due to the higher number of features.

Plot correlation between factors:

```{r}

plot_factor_cor(MOFAobject.trained)

```

Correlation between factors and metadata:

```{r, message=FALSE, warning=FALSE}

correlate_factors_with_covariates(MOFAobject.trained, 
                                  covariates = c("Day", "Horse", "Group", "Group_day"), 
                                  plot="log_pval",
                                  alpha="0.05" # p-value threshold
)


correlate_factors_with_covariates(MOFAobject.trained, 
  covariates = c("Day", "Horse", "Group", "Group_day"), 
  plot="r"
)

```

Factor 4 is significantly correlated with day, though factors 2-5 seem to have some correlation with day. Factor 4 is mostly explained by variance in the miRNA. Factor 2 is correlated with the 'group_day' variable, though its worth noting this factor had to be coerced into a numeric variable. For factor 2 miRNA and lncRNA explain more of the variance.

It may be worth looking at the proteomics weight for factor 1 and 5. In factor 1's case the variance is most strongly correlated with the proteomics but this may be because there is a much higher number of features. Factor 5 seems to have a slight correlation with the biological factors of interest.

```{r}

miRNA_f2 <- plot_top_weights(MOFAobject.trained, 
                  view = "miRNA", 
                  factor = 2, 
                  nfeatures = 10,
                  scale = F, 
                  abs = T
)

```

```{r}

lncRNA_f2 <- plot_top_weights(MOFAobject.trained, 
                  view = "lncRNA", 
                  factor = 2, 
                  nfeatures = 10,
                  scale = F, 
                  abs = T
)

```

```{r}

miRNA_f4 <- plot_top_weights(MOFAobject.trained, 
                  view = "miRNA", 
                  factor = 4, 
                  nfeatures = 10,
                  scale = F, 
                  abs = T
)

```

```{r}

proteomics_f3 <- plot_top_weights(MOFAobject.trained, 
                  view = "proteomics", 
                  factor = 3, 
                  nfeatures = 10,
                  scale = F, 
                  abs = T
)


# Annotate so Uniprot accessions are replaced by protein names
# The accessions are stored as factors so need to be converted back into character vectors
proteomics_f3$data$feature_id <- sf_annotation[as.character(proteomics_f3$data$feature_id),]

```

Plot miRNA factor 2, lncRNA factor 2, miRNA factor 4 and proteomics factor 3:

```{r, echo=FALSE}

ggarrange(miRNA_f2, lncRNA_f2, miRNA_f4)


proteomics_f3

```

Plot weights for factor 2:

```{r}

plot_weights(MOFAobject.trained,
 view = "miRNA",
 factor = 3,
 nfeatures = 20,     # Top number of features to highlight
 scale = F           # Scale weights from -1 to 1
)


plot_weights(MOFAobject.trained,
 view = "snRNA",
 factor = 3,
 nfeatures = 20,     # Top number of features to highlight
 scale = F           # Scale weights from -1 to 1
)



```

```{r}

# Add day as a factor to metadata for visualisation purposes
MOFAobject.trained@samples_metadata$Days <- as.factor(sf_pheno$Day)

# Palette for heatmap
heatmap_cols <- colorRampPalette(ArchRPalettes$coolwarm)(50)

# Plot heatmap
f5_sf_protein_heatmap <- plot_data_heatmap(MOFAobject.trained, 
        view = "proteomics",
        factor = 2,  
        features = 25,
        cluster_rows = TRUE, cluster_cols = TRUE,
        show_rownames = TRUE, show_colnames = FALSE,
        scale = "row",
        denoise = TRUE,
        color = heatmap_cols,
        border_color = NA,
        annotation_samples = c("Day", "Horse", "Group"),
        fontsize = 12
      )


f5_sf_protein_heatmap

```

Project the samples onto factor 2 & 4 for each time point.

```{r}

# Add day as a factor to metadata for visualisation purposes
MOFAobject.trained@samples_metadata$group_day <- as.factor(paste0(sf_pheno$group, "_", sf_pheno$Day))

```

```{r}

plot_factor(MOFAobject.trained, 
  factors = 3, 
  color_by = "group_day",
  add_violin = TRUE,
  dodge = TRUE
)

```

```{r}

plot_factor(MOFAobject.trained, 
  factors = 5, 
  color_by = "group_day",
  add_violin = TRUE,
  dodge = TRUE
)

```

The results were not saved for this model as a better performing version is explored below.

## Synovial fluid - version 2 (MEFISTO)

As the synovial fluid experimental design is more complex we will expand on the time and group variables in the MOFA2 model by making use of MEFISTO:

-   <https://raw.githack.com/bioFAM/MEFISTO_tutorials/master/MEFISTO_temporal.html>

-   <https://www.nature.com/articles/s41592-021-01343-9>

```{r}

colnames(sf_pheno)[c(5,6)] <- c("group", "sample")

# Create a variable for time
time <- t(sf_pheno$Day)
colnames(time) <- sf_pheno$sample
row.names(time) <- "time"

```

```{r}

sf_mofa_dat <- list(proteomics = as.matrix(sf_proteomics_data),
                        miRNA = as.matrix(sf_miRNA_data),
                        lncRNA = as.matrix(sf_lncRNA_data),
                        snRNA = as.matrix(sf_snRNA_data))

```

Create MOFAojbect:

```{r}

# Include group in the model 
MOFAobject <- create_mofa(sf_mofa_dat, groups = sf_pheno$group)

# Set time as a covariate in the model
MOFAobject <- set_covariates(MOFAobject, covariates = time)

print(MOFAobject)

```

Modify model parameters (including Mefisto options):

```{r}

# Scale views
data_opts <- get_default_data_options(MOFAobject)
data_opts$scale_views <- TRUE
head(data_opts)


# Set number of factors
model_opts <- get_default_model_options(MOFAobject)
model_opts$num_factors <- 6
head(model_opts)


# Set training options
train_opts <- get_default_training_options(MOFAobject)
train_opts$convergence_mode <- "slow"
train_opts$seed <- 123
train_opts$maxiter <- 3000
head(train_opts)


# Mefisto options
mefisto_opts <- get_default_mefisto_options(MOFAobject)
head(mefisto_opts)

```

Prepare MOFA object with modified parameters:

```{r, message=FALSE, warning=FALSE}

MOFAobject <- prepare_mofa(
  object = MOFAobject,
  data_options = data_opts,
  model_options = model_opts,
  training_options = train_opts,
  mefisto_options = mefisto_opts
)

```

Train model:

```{r, message=FALSE, warning=FALSE}

outfile <- file.path(getwd(),"sf_time_model.hdf5")
MOFAobject.trained <- run_mofa(MOFAobject, outfile, use_basilisk=TRUE)

```

Add phenotype data to the model:

```{r}

samples_metadata(MOFAobject.trained) <- sf_pheno

```

Visualise variance explained:

```{r}

plot_variance_explained(
  MOFAobject.trained, 
  x = "view", 
  y = "factor", 
  plot_total = T
)


```

Interestingly the miRNA contributes more strongly to the variance in the control samples. The MOFA model has learnt scale parameters for each factor, which gives an indication of the smoothness per factor along the covariate (time) and are between 0 and 1. A scale of 0 means that the factor captures variation independent of time, a value close to 1 tells us that this factor varys very smoothly along time.

```{r}

get_scales(MOFAobject.trained)

```

Here it can be seen factor 3 varies smoothly along time.

Plot correlation between factors:

```{r}

plot_factor_cor(MOFAobject.trained)

```

Correlation between factors and metadata:

```{r, message=FALSE, warning=FALSE}

correlate_factors_with_covariates(MOFAobject.trained, 
                                  covariates = c("Day", "Horse"), 
                                  plot="log_pval",
                                  alpha="0.15" # p-value threshold
)


correlate_factors_with_covariates(MOFAobject.trained, 
  covariates = c("Day", "Horse"), 
  plot="r"
)

```

Factor 3 is most strongly correlated with day, followed by 5. For factor 3 most of the variation comes from the miRNA, followed by the snRNA. For factor 5 most of the variation comes from the proteomics data.

```{r}

miRNA_f3 <- plot_top_weights(MOFAobject.trained, 
                  view = "miRNA", 
                  factor = 3, 
                  nfeatures = 10,
                  scale = F, 
                  abs = T
)

```

```{r}

snRNA_f3 <- plot_top_weights(MOFAobject.trained, 
                  view = "snRNA", 
                  factor = 3, 
                  nfeatures = 10,
                  scale = F, 
                  abs = T
)

```

```{r}

proteomics_f5 <- plot_top_weights(MOFAobject.trained, 
                  view = "proteomics", 
                  factor = 5, 
                  nfeatures = 10,
                  scale = F, 
                  abs = T
)


# Annotate so Uniprot accessions are replaced by protein names
# The accessions are stored as factors so need to be converted back into character vectors
proteomics_f5$data$feature_id <- sf_annotation[as.character(proteomics_f5$data$feature_id),]

```

Plot miRNA factor 2, lncRNA factor 2, miRNA factor 4 and proteomics factor 3:

```{r, echo=FALSE}

ggarrange(miRNA_f3, snRNA_f3)


proteomics_f5

```

Plot weights for factor 3:

```{r}

plot_weights(MOFAobject.trained,
 view = "miRNA",
 factor = 3,
 nfeatures = 20,     # Top number of features to highlight
 scale = F           # Scale weights from -1 to 1
)


plot_weights(MOFAobject.trained,
 view = "lncRNA",
 factor = 3,
 nfeatures = 20,     # Top number of features to highlight
 scale = F           # Scale weights from -1 to 1
)



```

Heatmaps for factors...

Factor 5, protein:

```{r}

# Add day as a factor to metadata for visualisation purposes
MOFAobject.trained@samples_metadata$Days <- as.factor(plasma_pheno$Day)

# Palette for heatmap
heatmap_cols <- colorRampPalette(ArchRPalettes$coolwarm)(50)

# Plot heatmap
f5_sf_protein_heatmap <- plot_data_heatmap(MOFAobject.trained, 
  view = "proteomics",
  factor = 5,  
  features = 25,
  cluster_rows = TRUE, cluster_cols = TRUE,
  show_rownames = TRUE, show_colnames = FALSE,
  scale = "row",
  denoise = TRUE,
  color = heatmap_cols,
  border_color = NA,
  annotation_samples = c("Day", "Horse", "group"),
  fontsize = 12
)


f5_sf_protein_heatmap 

```

Factor 2, miRNA:

```{r}

# Plot heatmap
f2_sf_miRNA_heatmap <- plot_data_heatmap(MOFAobject.trained, 
  view = "miRNA",
  factor = 2,  
  features = 25,
  cluster_rows = TRUE, cluster_cols = TRUE,
  show_rownames = TRUE, show_colnames = FALSE,
  scale = "row",
  denoise = TRUE,
  color = heatmap_cols,
  border_color = NA,
  annotation_samples = c("Day", "Horse", "group"),
  fontsize = 12
)


f2_sf_miRNA_heatmap 

```

Factor 2, protein:

```{r}

# Plot heatmap
f2_sf_protein_heatmap <- plot_data_heatmap(MOFAobject.trained, 
  view = "proteomics",
  factor = 2,  
  features = 25,
  cluster_rows = TRUE, cluster_cols = TRUE,
  show_rownames = TRUE, show_colnames = FALSE,
  scale = "row",
  denoise = TRUE,
  color = heatmap_cols,
  border_color = NA,
  annotation_samples = c("Day", "Horse", "group"),
  fontsize = 12
)


f2_sf_protein_heatmap 

```

Factor 3, miRNA:

```{r}

# Plot heatmap
f3_sf_miRNA_heatmap <- plot_data_heatmap(MOFAobject.trained, 
  view = "miRNA",
  factor = 3,  
  features = 25,
  cluster_rows = TRUE, cluster_cols = TRUE,
  show_rownames = TRUE, show_colnames = FALSE,
  scale = "row",
  denoise = TRUE,
  color = heatmap_cols,
  border_color = NA,
  annotation_samples = c("Day", "Horse", "group"),
  fontsize = 12
)


f3_sf_miRNA_heatmap 

```

Factor 3, protein:

```{r}

# Plot heatmap
f3_sf_protein_heatmap <- plot_data_heatmap(MOFAobject.trained, 
  view = "proteomics",
  factor = 3,  
  features = 25,
  cluster_rows = TRUE, cluster_cols = TRUE,
  show_rownames = TRUE, show_colnames = FALSE,
  scale = "row",
  denoise = TRUE,
  color = heatmap_cols,
  border_color = NA,
  annotation_samples = c("Day", "Horse", "group"),
  fontsize = 12
)


f3_sf_protein_heatmap 

```

Project the samples onto factor 5 for each time point.

```{r}

# Add day as a factor to metadata for visualisation purposes
MOFAobject.trained@samples_metadata$group_day <- as.factor(paste0(sf_pheno$group, "_", sf_pheno$Day))

```

```{r}

plot_factor(MOFAobject.trained, 
  factors = 5, 
  color_by = "group_day",
  add_violin = TRUE,
  dodge = TRUE
)

```

Save results:

```{r}

sf_res <- get_weights(
    MOFAobject.trained,
    views = "all",
    factors = "all",
    abs = FALSE,
    scale = FALSE,
    as.data.frame = FALSE
)



sf_res_df <- get_weights(
    MOFAobject.trained,
    views = "all",
    factors = "all",
    abs = FALSE,
    scale = FALSE,
    as.data.frame = TRUE
)


```

```{r}

saveRDS(sf_res, "./data/mofa2_sf.rds")
saveRDS(sf_res_df, "./data/mofa2_sf_df.rds")

```

# Summary/conclusions

Joint dimensionality reduction identified factors associated with time after OA induction in synovial fluid and plasma. If more time was available I would have also compared/contrasted other multi-omics methods such as JIVE and iCluster.

<!-- Below is some code that will be hidden in the outputs to save figures. -->

```{r, echo=FALSE, eval=FALSE, fig.show='hide', message=FALSE}


variance <- plot_variance_explained(
  MOFAobject.trained, 
  x = "view", 
  y = "factor", 
  plot_total = T
)


correlation <- correlate_factors_with_covariates(MOFAobject.trained, 
  covariates = c("Day", "Horse"), 
  plot="log_pval", 
  alpha="0.15",
  color = heatmap_cols,
  border_color = NA,
  fontsize = 12
)



# Plot heatmap
f2_plasma_protein_heatmap <- plot_data_heatmap(MOFAobject.trained, 
  view = "proteomics",
  factor = 2,  
  features = 25,
  cluster_rows = TRUE, cluster_cols = TRUE,
  show_rownames = TRUE, show_colnames = FALSE,
  scale = "row",
  denoise = TRUE,
  color = heatmap_cols,
  border_color = NA,
  annotation_samples = c("Day", "Horse"),
  fontsize = 12,
  legend = FALSE
)



fig1 <- ggarrange(f2_plasma_protein_heatmap[[4]], ggarrange(variance[[1]], correlation[[4]],                        
                    nrow = 2, 
                    ncol = 1,
                    labels = c("B", "C")),
                  
          nrow = 1,
          ncol = 2,
          labels = c("A"))



# Group figure
ggsave(
  filename = "plasma_mofa.PDF",
  plot = fig1,
  width = 8,
  height = 7,
  units = c("in"),
  dpi = 300,
  limitsize = TRUE,
  bg = "white"
)


# Correlation
ggsave(
  filename = "plasma_factor_group_correlation_mofa.PDF",
  plot = correlation[[4]],
  width = 4,
  height = 3.5,
  units = c("in"),
  dpi = 300,
  limitsize = TRUE,
  bg = "white"
)



# Variance 
ggsave(
  filename = "plasma_factor_variation_mofa.PDF",
  plot = variance[[1]],
  width = 4.5,
  height = 3.5,
  units = c("in"),
  dpi = 300,
  limitsize = TRUE,
  bg = "white"
)


```

```{r, echo=FALSE, eval=FALSE, fig.show='hide', message=FALSE}

ggsave(
  filename = "protein_plasma_factor2.PDF",
  plot = f2_plasma_protein_heatmap[[4]],
  width = 6,
  height = 6,
  units = c("in"),
  dpi = 300,
  limitsize = TRUE,
  bg = "white"
)



ggsave(
  filename = "miRNA_plasma_factor2.PDF",
  plot = f2_plasma_miRNA_heatmap[[4]],
  width = 6,
  height = 6,
  units = c("in"),
  dpi = 300,
  limitsize = TRUE,
  bg = "white"
)


```

```{r, echo=FALSE, eval=FALSE, fig.show='hide', message=FALSE}

ggsave(
  filename = "protein_plasma_factor2.PDF",
  plot = f2_plasma_protein_heatmap[[4]],
  width = 6,
  height = 6,
  units = c("in"),
  dpi = 300,
  limitsize = TRUE,
  bg = "white"
)



ggsave(
  filename = "miRNA_plasma_factor2.PDF",
  plot = f2_plasma_miRNA_heatmap[[4]],
  width = 6,
  height = 6,
  units = c("in"),
  dpi = 300,
  limitsize = TRUE,
  bg = "white"
)


```

# References
